+++ price-feeder/oracle/oracle_test.go
@@ -15,8 +15,6 @@ import (
 	"github.com/umee-network/umee/price-feeder/oracle/client"
 	"github.com/umee-network/umee/price-feeder/oracle/provider"
 	"github.com/umee-network/umee/price-feeder/oracle/types"
-
-	oracletypes "github.com/umee-network/umee/v2/x/oracle/types"
 )
 
 type mockProvider struct {
@@ -109,6 +107,7 @@ func (ots *OracleTestSuite) SetupSuite() {
 		},
 		time.Millisecond*100,
 		make(map[string]sdk.Dec),
+		make(map[string]config.ProviderEndpoint),
 	)
 }
 
@@ -156,15 +155,8 @@ func (ots *OracleTestSuite) TestPrices() {
 			},
 		},
 	}
-	acceptList := oracletypes.DenomList{
-		oracletypes.Denom{
-			BaseDenom:   "UUMEE",
-			SymbolDenom: "UMEE",
-			Exponent:    6,
-		},
-	}
 
-	ots.Require().Error(ots.oracle.SetPrices(context.TODO(), acceptList))
+	ots.Require().Error(ots.oracle.SetPrices(context.TODO()))
 	ots.Require().Empty(ots.oracle.GetPrices())
 
 	// use a mock provider without a conversion rate for these stablecoins
@@ -187,7 +179,7 @@ func (ots *OracleTestSuite) TestPrices() {
 		},
 	}
 
-	ots.Require().Error(ots.oracle.SetPrices(context.TODO(), acceptList))
+	ots.Require().Error(ots.oracle.SetPrices(context.TODO()))
 
 	prices := ots.oracle.GetPrices()
 	ots.Require().Len(prices, 0)
@@ -236,7 +228,7 @@ func (ots *OracleTestSuite) TestPrices() {
 		},
 	}
 
-	ots.Require().NoError(ots.oracle.SetPrices(context.TODO(), acceptList))
+	ots.Require().NoError(ots.oracle.SetPrices(context.TODO()))
 
 	prices = ots.oracle.GetPrices()
 	ots.Require().Len(prices, 4)
@@ -289,7 +281,7 @@ func (ots *OracleTestSuite) TestPrices() {
 		},
 	}
 
-	ots.Require().NoError(ots.oracle.SetPrices(context.TODO(), acceptList))
+	ots.Require().NoError(ots.oracle.SetPrices(context.TODO()))
 	prices = ots.oracle.GetPrices()
 	ots.Require().Len(prices, 4)
 	ots.Require().Equal(sdk.MustNewDecFromStr("3.70"), prices["UMEE"])
@@ -341,7 +333,7 @@ func (ots *OracleTestSuite) TestPrices() {
 		},
 	}
 
-	ots.Require().NoError(ots.oracle.SetPrices(context.TODO(), acceptList))
+	ots.Require().NoError(ots.oracle.SetPrices(context.TODO()))
 	prices = ots.oracle.GetPrices()
 	ots.Require().Len(prices, 4)
 	ots.Require().Equal(sdk.MustNewDecFromStr("3.71"), prices["UMEE"])