+++ x/leverage/simulation/operations.go
@@ -356,6 +356,7 @@ func randomWithdrawFields(
 
 	uTokens := getSpendableUTokens(ctx, acc.Address, bk, lk)
 	uTokens = uTokens.Add(lk.GetBorrowerCollateral(ctx, acc.Address)...)
+
 	uTokens = simtypes.RandSubsetCoins(r, uTokens)
 
 	if uTokens.Empty() {
@@ -419,6 +420,7 @@ func randomLiquidateFields(
 	}
 
 	borrowed := lk.GetBorrowerBorrows(ctx, borrower.Address)
+
 	borrowed = simtypes.RandSubsetCoins(r, borrowed)
 	if borrowed.Empty() {
 		return liquidator, borrower, sdk.Coin{}, "", true